+++ packages/collections/test/mock/server.test.js
@@ -41,12 +41,28 @@ describe('GET', () => {
     expect(response.statusCode).to.equal(403);
   });
 
-  it('should consume all results as a stream', async () => {
+  it('/collection/name/key should return a single item', async () => {
     api.createCollection('my-collection');
 
     api.upsert('my-collection', 'x', { id: 'x' });
-    api.upsert('my-collection', 'y', { id: 'y' });
-    api.upsert('my-collection', 'z', { id: 'z' });
+
+    const response = await request({
+      method: 'GET',
+      path: 'collections/my-collection/x',
+    });
+
+    const item = await response.body.json();
+
+    expect(response.statusCode).to.eql(200);
+    expect(item).to.eql({ key: 'x', value: { id: 'x' } });
+  });
+
+  it('/collection/name should stream all results', async () => {
+    api.createCollection('my-collection');
+
+    api.upsert('my-collection', 'x', 'xx');
+    api.upsert('my-collection', 'y', 'yy');
+    api.upsert('my-collection', 'z', 'zz');
 
     const response = await request({
       method: 'GET',
@@ -57,46 +73,28 @@ describe('GET', () => {
     await streamResponse(response, item => results.push(item));
 
     expect(results).to.eql([
-      { key: 'x', value: { id: 'x' } },
-      { key: 'y', value: { id: 'y' } },
-      { key: 'z', value: { id: 'z' } },
+      { key: 'x', value: 'xx' },
+      { key: 'y', value: 'yy' },
+      { key: 'z', value: 'zz' },
     ]);
   });
 
-  it('should consume a single result as a stream', async () => {
+  it('/collection/name?key=* should stream some results', async () => {
     api.createCollection('my-collection');
 
-    api.upsert('my-collection', 'x', { id: 'x' });
-    api.upsert('my-collection', 'y', { id: 'y' });
-    api.upsert('my-collection', 'z', { id: 'z' });
+    api.upsert('my-collection', 'ax', 'x');
+    api.upsert('my-collection', 'ay', 'y');
+    api.upsert('my-collection', 'az', 'z');
 
     const response = await request({
       method: 'GET',
-      path: 'collections/my-collection/y',
+      path: 'collections/my-collection?key=*z',
     });
     const results = [];
 
     await streamResponse(response, item => results.push(item));
 
-    expect(results).to.eql([{ key: 'y', value: { id: 'y' } }]);
-  });
-
-  it('should consume results as a stream with a wildcard', async () => {
-    api.createCollection('my-collection');
-
-    api.upsert('my-collection', 'ax', { id: 'x' });
-    api.upsert('my-collection', 'ay', { id: 'y' });
-    api.upsert('my-collection', 'az', { id: 'z' });
-
-    const response = await request({
-      method: 'GET',
-      path: 'collections/my-collection/*z',
-    });
-    const results = [];
-
-    await streamResponse(response, item => results.push(item));
-
-    expect(results).to.eql([{ key: 'az', value: { id: 'z' } }]);
+    expect(results).to.eql([{ key: 'az', value: 'z' }]);
   });
 });